move dialog to own directory
