+++ src/ui/pages/Settings.tsx
@@ -15,7 +15,7 @@ import { ReactNode, useContext, useEffect, useState } from 'react'
 import { Route, Routes } from 'react-router'
 import { R } from 'src/core/util/functional'
 import { stringSorter } from 'src/core/util/sort'
-import { AlertDialog } from '../components/AlertDialog'
+import { AlertDialog } from '../components/dialog/AlertDialog'
 import SideTabs from '../components/side-tabs/SideTabs'
 import useDisplayError from '../hooks/displayError'
 import { usePathSegment } from '../hooks/routing'
