+++ src/ui/top-level/ErrorMessageModal.tsx
@@ -2,7 +2,7 @@ import { ErrorIcon } from '@openhome-ui/components/Icons'
 import { ErrorContext } from '@openhome-ui/state/error'
 import { Callout } from '@radix-ui/themes'
 import { useContext } from 'react'
-import { AlertDialog } from '../components/AlertDialog'
+import { AlertDialog } from '@openhome-ui/components/dialog/AlertDialog'
 
 export default function ErrorMessageModal() {
   const [errorState, dispatchErrorState] = useContext(ErrorContext)
